Update UnixSshFileSystemProvider.java
Not all versions of stat support creation time. This will handle the case where it returns something other than a number. I've tried with 3 different versions of stat and 2 of them returned something other than a number. In one case it was "W" and in another it was "?". The spec for BasicFileAttributes says that the epoche is a valid default value in case the file system doesn't support the feature.
Instead of dealing with exceptions, you can validate that it's a number using some other means.